fix(mongodb-operator): patch quirk in operator
The operator is weird and requires that every namespace that wants to deploy
MongoDB resources have a service account. This is a bit of a pain, but we
can copy over the service account from the operator's namespace.

https://github.com/mongodb/mongodb-kubernetes-operator/issues/850

We do some Nix magic here. Essentially, we look for all namespaces that have
MongoDBCommunity resources and copy over the service account from the
operator's namespace. We can do this by defining a submodule that Nix will
merge with the existing `namespaces` submodule provided by transpire.